Fix field precision specifier
ldap_mapper.c: In function `ldap_build_partial_cert_filter':
ldap_mapper.c:731:3: warning: field precision specifier `.*' expects argument of type `int', but argument 5 has type `long int' [-Wformat=]
   DBG2("ldap_build_cert_filter(): unrecognized certificate "
   ^
ldap_mapper.c:736:3: warning: field precision specifier `.*' expects argument of type `int', but argument 5 has type `long int' [-Wformat=]
   DBG2("ldap_build_cert_filter(): no values for certificate "
   ^
ldap_mapper.c:740:2: warning: field precision specifier `.*' expects argument of type `int', but argument 5 has type `long int' [-Wformat=]
  DBG4("ldap_build_cert_filter(): building subfilter '%.*s'='%.*s'",
  ^
ldap_mapper.c:740:2: warning: field precision specifier `.*' expects argument of type `int', but argument 7 has type `long int' [-Wformat=]

+++ src/mappers/ldap_mapper.c
@@ -729,16 +729,16 @@ ldap_build_partial_cert_filter(const char *map, X509 *x509)
 		values = certs;
 	} else {
 		DBG2("ldap_build_cert_filter(): unrecognized certificate "
-		     "attribute '%.*s'", q - p - 1, p + 1);
+		     "attribute '%.*s'", (int)(q - p - 1), p + 1);
 		return NULL;
 	}
 	if (values == NULL) {
 		DBG2("ldap_build_cert_filter(): no values for certificate "
-		     "attribute '%.*s'", q - p - 1, p + 1);
+		     "attribute '%.*s'", (int)(q - p - 1), p + 1);
 		return NULL;
 	}
 	DBG4("ldap_build_cert_filter(): building subfilter '%.*s'='%.*s'",
-	     p - map, map, q - p - 1, p + 1);
+	     (int)(p - map), map, (int)(q - p - 1), p + 1);
 	for (n = 0, buf_len = 0; values[n] != NULL; n++) {
 		buf_len++;
 		buf_len += (p - map);
